integ tests: move reusable benchmarks functions to common module
